complx: Print full version string in help output
In the output for `complx -h', print the full version number (v4.15.4),
not only the major and minor version (v4.15).

Although the full version number is clearly visible when opening the
complx GUI (either in the title or the Help > About page), this commit
prevents confusion for users who run `complx -h' instead.