Added missing const marker on lc3_subroutine_call_info_cmp operator method.
The Clang compiler on Mac was failing due to an error with this line where the operator method needed to be marked const but wasn’t. I have tested complx and the tests on recent homeworks after compiling with this modification and everything seems to be working just fine now.