complx: Accept config file path on command line
Apparently wxWidgets tries to delete the config file before saving,
which makes simply bind-mounting C:\some\path\.Complx to ~/.Complx as a
volume infeasible. So instead, mount a directory containing the config
file and tell Complx to read it from there. To make this dream happen,
complx needs to be able to optionally accept a path to a config file on
the command line.

Strangely enough relative paths are relative to the user's homedir, at
least on Linux.